+++ playground/samples/filter/duotone/sample.js
@@ -20,6 +20,7 @@ lut = lut.colourspace(vips.Interpretation.srgb/* 'srgb' */, {
 // Image source: https://www.flickr.com/photos/jasonidzerda/3987784466
 let im = vips.Image.newFromFile('owl.jpg');
 // let im = vips.Image.newFromFile('owl.tif');
+// let im = vips.Image.newFromFile('owl.jxl');
 // let im = vips.Image.newFromFile('transparency_demo.png');
 // let im = vips.Image.newFromFile('banana.webp', { n: -1 });
 // let im = vips.Image.newFromFile('banana.gif', { n: -1 });